requirements.txt
psycopg2-binary
flask
flask-cors
prometheus-flask-exporter
python-dotenv
